feat(writer): retry when detecting a closed connection
Currently, when a writer has low volume, the server may close the
connection between messages, causing an EOF error when attempting to
send again. This case is easy to test for and retry internally, which is
being done here.

This particularly is useful for async writers that only spend sparse
data, preventing messages from being lost due to this normal behavior of
client/server interactions.