writer: make the force flush timer accurate
Prior, every writer to a partition had its own timer that ticked at
1/10th the BatchTimeout interval. Each tick, the code would check
whether BatchTimeout has elapsed since the prior flush and, if so,
trigger a flush.

This change makes the timer more accurate so that it only needs to tick
once every BatchTimeout interval, and only needs to begin a timeout
on the first message for a batch.