Handle end of fetched batches of messages gracefully
Previously, the batch read functions returned io.EOF to signal the
end of the batch.  However, this caused ambiguity because there are
valid io.EOF errors from other sources (e.g. a connection that is
closed by the broker).  Because io.EOF can leave the conection in
an unusable state, we are closing the connection and logging it.
However, this also means that we're closing the connection after
every single fetch request, which is not desirable.

This PR also fixes an accidental line deletion of deadline
management that was introduced in #277.

Fixes #258