Don't close connection when end of batch is read correctly
io.EOF is used as a marker that a batch of messages has been
completely read.  The batch will repackage an io.EOF from the
underlying connection as an io.ErrUnexpectedEOF.

The previous behavior closed the connection on an io.EOF and
restarted the reader loop, which will be much less efficient than
continuing to use the connnection.