Fix zstd decoder leak (#446)
TL;DR: using `sync.Pool` to pool encoders/decoders that are not eligible for
deallocation once the pool loses their reference leads to leaks.

The docs from `sync.Pool` state:
> Any item stored in the Pool may be removed automatically at any time without
> notification. If the Pool holds the only reference when this happens, the
> item might be deallocated.

`zstd.Decoder`s spin up their own goroutines, which hold a strong ref to owning
decoder. When `decPool` downsizes (usually on GC runs), an arbitrary number of
decoders are evicted from the pool. Since their goroutines are still running
(observed: waiting on `*blockDec.startDecoder()`, at instruction
`for range b.input { ... }`) and thus holding references, these decoders are not
eligible for deallocation — instead becoming unreachable. As new decoders are
created and evicted, the number of running goroutines continuously increases,
eventually causing the program to run out of memory and/or peg the CPU.

I originally noticed this under heavy concurrent load, but managed to create a
trivial, serial test case where the problem is demonstrated.

This change introduces a pretty naive fix. A more robust solution should be
discussed and implemented (or at least a more configurable one). The goal here
was to minimize changes to existing code in order to help focus on the existing
issue and foster discussion.

Likewise, there isn't much value to the test case other than demonstrating the
original problem, and should probably be dropped once a robust solution is put
in place to prevent this leak.