ReadBatchWith bug fix and new MaxWaitTime option (#482)
This fixes a bug where ReadBatchWith could incorrectly return `io.EOF`
in certain `errShortRead` scenarios.  Since `io.EOF` means the batch
was handled correctly, this could cause consumers to continue to use
a connection that is in a bad state.

This also adds a MaxWait option that can be used to improve the
way that the conn deadline and the fetch request interact.  Instead
of trying to calculate a correct timeout and change the max wait
time, this allows the caller to explicitly configure both the wait
and the deadline independently of one another.