Fix Endless waiting for consumer connection (#227) (#430)
When the high-level Reader API is used, e.g. when calling
Reader.FetchMessage, and the Reader is configured with a non-empty
GroupID, make sure that the call returns an error when the call to
ConsumerGroup.Next fails. This can happen, for example, when the Kafka
server isn't reachable.

The call will fail only after Reader.config.MaxAttempts attempts
calling ConsumerGroup.Next were performed internally by Reader.run
and failed.

Previously, Reader.FetchMessage blocked even when the Kafka cluster
could not possibly be reached, making it impractical to distinguish
between not getting messages because there are none and not getting
messages because we cannot connect to the server.

Two new test cases are provided (one of them tests for behavior which
was working correctly, namely that Reader without a GroupID will return
an error when it cannot connect -- that was added for completeness).

Co-authored-by: David Čepelík <david.cepelik@showmax.com>