feat: first draft of the reader group
Reader groups are a way of consuming messages from multiple partitions
without consumer groups. This is useful, e.g. when you are using
compacted topics and don't replicate the state from Kafka. Then you can
use these reader groups for consuming all messages to build a projection
of the data in memory or in an ephemeral storage

Relates to #386