allow resetting offsets to either earliest or latest
Kafka has a notion of magic sentinel numbers (-1, -2) for consumer offset values [1],
which are meant to signal that the consumer should start from earliest (first) or
latest (last) available offset for a given partition.

Instead of obeying that signal, we would unconditionally respect the default
`StartOffset` value from the configuration when reading a negative offset upon
synchronizing consumergroup state (such as at initialization).

Hence you could *only* have either "read from earliest" or "read from latest".
behavior from your consumers when you reset the offsets to one of these
sentinel values.

With this change, the consumer will respect the returned value if it is -1
or -2, allowing an operator to start a consumer at earliest or latest irrespective
of the configured value of `StartOffset`.

[1] From the kafka wiki for the [ListOffset API](https://cwiki.apache.org/confluence/display/KAFKA/A+Guide+To+The+Kafka+Protocol#AGuideToTheKafkaProtocol-OffsetAPI(AKAListOffset)):

> Field: Time
>
> Description:
> Used to ask for all messages before a certain time (ms). There are two special values
> Specify -1 to receive the latest offset (i.e. the offset of the next coming message)
> and -2 to receive the earliest available offset. This applies to all versions of the API.
> Note that because offsets are pulled in descending order, asking for the earliest offset
> will always return you a single element.