Remove existing message set reader.
The existing reader checked the magic byte of the first message
set in the response, and processed subsequent sets assuming
that version. Because it delegated to the v1 and v2 set readers
based on this version, we end up having two separate set readers,
effectively.

A following commit will create a new message set reader to replace
it that unifies the reader logic and parses the magic byte in
each message set in a response.