Fix bug when seeking offsets in compressed batches
The reader was failing to mark the current message as read when
fast forwarding through a compressed message set with more than
one inner messages. Additionally, it was only discarding the
message key instead of both the key and the value.

After fixing this, the reader was able to discard the message
correctly, and start parsing the next message header.