Fix message builder compression code.
Due to a bug in the builder code, each compressed message set
was creating a compressed message set for each message. This
fixes that behavior so that the outer message composes N
inner messages.

The offset behavior between the v0 and v1 message set types
is also accounted for.

See: https://kafka.apache.org/documentation/#messages