Allow the resolver to override the port setting
The custom resolver couldn't override the default port even if id had been explicitly provided.
The reason for that is unclear as it contradicts the open closed principle.
Why: Common address resolvers should alway only return the host. If a port is being provided this has been intended.

Purpose of this Patch:
Access a kafka cluster behind a firewall via portforwarding on a jump host.
jumphost: 9095 -> kafkahost1:9092, jumphost: 9096 -> kafkahost2:9092, ...