Fix go vet issues: example tests, returning a mutex copy (#783)
Example tests are renamed to fit required naming conventions. This
caused them to be relocated in generated docs.

Also, a problem with writer.newBatchQueue is fixed; in order to avoid an
unnecessary copy of a newly created object holding a mutex, the mutex
field within batchQueue is now a pointer, so that copying on return
doesn't copy the mutex itself. The same treatment is needed for the cond
field; the same no-copy rule applies to it.