don't attempt to read messages when offset is greater than hwm
Based on
> The high watermark offset is the offset of the last message that was
successfully copied to all of the log’s replicas.

the high watermark is the last message that was copied to all of the
logs replicas and consumers should not read past that message.

Previously we stopped reading messages when offset == highWaterMark
which led to messages being skipped when the offset was equal to the
highWaterMark.