add finalizer to Reader
we run into issues where users don't call Close() on their readers
before creating new ones. This causes issues with the consumer groups
as more members are added to the group which can exceed the number of
partitions and stop work. This should prevent that from happening in
some cases by setting a finalizer which closes the Reader