Don't invoce r.log() if not r.debug
Profiling shows that up to 40% of our allocations happen in
r.markRead(): just to allocate a slice to pass the "remain" arg, which
will be discarded because we don't run with r.debug=true.

This makes the code slightly verbose, having to check the debug mode
before invoking r.log(), but it's a huge difference in high-throughput
applications.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>